// frontend/src/client/types.gen.ts
export type DocumentPublic = {
    title: string;
    content: string;
    id: string;
    accuracy_report?: (string | null);
    is_accurate?: (boolean | null);
    created_at: string;
    owner_id: string;
};